FILMZ-1 Fixes the app config setup
Cookie configs now work as expected

Adds pytest module for Web MVC testing (integration testing)

Cleaned up dashboard.html for further changes

Used ChatGPT to help with this. Flask sets up the PYTHONPATH for you, but pytest does not. This led to errors about the imports. So we had to include the PYTHONPATH environment variable in the test docker-compose file.