FILMZ-7 Adds UI and endpoints so users can select their active team
A lot of changes in this one:
- Fix bugs with Tailwind and JS packing due to dockerfile changes. Had to override entrypoint
- Adds user team list end point and list UI
- Changes the way user profiles are cached. It now includes the team details so Jinja can include the logo and name when necessary

Using a loop with Jinja context and JS functions can be annoying.
I had to essentially loop through the teams and generate JS functions. This isn't ideal. For a small page this is okay.

But it seems like it would be best to use fetch() to get the teams data. That way alpine can deal with the data securely instead of relying on inline JS event handlers